fonction combat fini+menu
